+++ CHANGELOG.md
@@ -9,3 +9,15 @@
 ## [1.0.1] - 24/06/2020
 
 - Added example and formated files
+
+## [1.0.2] - 26/06/2020
+
+- Added prepay cards!
+- Fixed bugs on card
+- Fixed width and height display resolutions
+- Easy way to decide what type of card you need
+- Soon Paypal
+
+## [1.0.3] - 28/06/2020
+
+- Added english language support